+++ inst/shinyMobile-1.0.0.9000/js/shinyMobile.js
@@ -77,7 +77,7 @@ $((function() {
         var reconnectToast = app.toast.create({
             icon: '<i class="icon f7-icons">bolt_fill</i>',
             position: "center",
-            text: 'Oups... disconnected </br> </br> <div class="row"><button onclick="Shiny.shinyapp.reconnect();" class="toast-button button color-green col">Reconnect</button><button onclick="location.reload();" class="toast-button button color-red col">Reload</button></div>'
+            text: 'Disconnected ... </br> </br> <div class="row"><button onclick="Shiny.shinyapp.reconnect();" class="toast-button button color-green col">Reconnect</button><button onclick="location.reload();" class="toast-button button color-red col">Reload</button></div>'
         }).open();
         $(".toast-button").on("click", (function() {
             reconnectToast.close();
@@ -98,10 +98,31 @@ $((function() {
             $(".toolbar").css("margin-bottom", "20px");
         }
     }
-    $(".tabs-standalone").css("height", "auto");
     if (app.params.filled && app.params.dark && $("body").attr("class") !== "#ffffff") {
         $(".demo-send-message-link").find("i").addClass("color-white");
     }
+    const skeletonClass = "skeleton-text skeleton-effect-fade";
+    $(document).on("shiny:outputinvalidated", (function(e) {
+        $("#" + e.target.id).addClass(skeletonClass);
+    }));
+    $(document).on("shiny:value", (function(e) {
+        $("#" + e.target.id).removeClass(skeletonClass);
+    }));
+    window.ran = false;
+    const skeletonTargets = [ ".page-content", ".navbar", ".toolbar" ];
+    $(document).on("shiny:connected", (function(event) {
+        for (target of skeletonTargets) {
+            $(target).addClass(skeletonClass);
+        }
+    }));
+    $(document).on("shiny:idle", (function(event) {
+        if (!window.ran) {
+            for (target of skeletonTargets) {
+                $(target).removeClass(skeletonClass);
+            }
+        }
+        window.ran = true;
+    }));
     isSplitLayout = $("#app").hasClass("split-layout");
     if (app.params.dark) {
         $("body").addClass("dark");
@@ -111,15 +132,6 @@ $((function() {
             }
         }
         $(".appbar").addClass("theme-dark");
-        $(".demo-facebook-card .card-footer").css("background-color", "#1c1c1d");
-        $(".fab-label").css("background-color", "var(--f7-fab-label-text-color)");
-        $(".fab-label").css("color", "var(--f7-fab-text-color)");
-        $(".accordion-item .item-content .item-inner").css("color", "white");
-        $(".accordion-item .accordion-item-content").css("color", "white");
-        var sidebarPanel = $("#f7-sidebar-view").find(".page-content");
-        $(sidebarPanel).css("background-color", "#1e1e1e");
-        var sidebarItems = $("#f7-sidebar-view").find("li");
-        $(sidebarItems).css("background-color", "#171717");
     } else {
         $("body").addClass("light");
         if (isSplitLayout) {